% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWords.R
\name{filterWords}
\alias{filterWords}
\title{Filters a vector of words to ones compatible with a}
\usage{
filterWords(x, words = NULL)
}
\arguments{
\item{x}{a woRdleStatus}

\item{words}{either a vector of words or internal list will be used}
}
\value{

}
\description{
Filters a vector of words to ones compatible with a
}
