% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterStats.R
\name{letterStats}
\alias{letterStats}
\title{letterStats}
\usage{
letterStats(words)
}
\arguments{
\item{words}{a vector of words}
}
\value{
dataframe containing statistics on frequency of letters in each word
}
\description{
letterStats
}
\examples{
letterStats(words = c("APPLE","PROSE","PUPPY"))

}
