% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woRdle.R
\name{plot.woRdle}
\alias{plot.woRdle}
\title{Plotting function for woRdle S3 object}
\usage{
\method{plot}{woRdle}(x, ...)
}
\arguments{
\item{x}{a woRdle object}

\item{...}{unused. compliance with S3}
}
\value{
a plot
}
\description{
Plotting function for woRdle S3 object
}
