% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woRdleGuess.R
\name{plot.woRdleGuess}
\alias{plot.woRdleGuess}
\title{plot a guess}
\usage{
\method{plot}{woRdleGuess}(x, ...)
}
\arguments{
\item{x}{a guess object}

\item{...}{unused}
}
\value{

}
\description{
plot a guess
}
