% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatewoRdleStatus.R
\name{updatewoRdleStatus}
\alias{updatewoRdleStatus}
\title{Updates a woRdleStatus with a woRdleGuess}
\usage{
updatewoRdleStatus(x, guess)
}
\arguments{
\item{x}{a woRdleStatus}

\item{guess}{a woRdleGuess}
}
\value{
a woRdleStatus
}
\description{
Updates a woRdleStatus with a woRdleGuess
}
\examples{
game_0 <- woRdleStatus()
guess_1 <- woRdleGuess("APPLE", "YGG00")
game_1 <- updatewoRdleStatus(game_0, guess_1)
}
